Remove revision specification for dn-common-utils-dev.
There was a discrepancy between the the base package dependency on
dn-common-utils, which did not use a revision specification, and the
dev package dependency on dn-common-utils-dev which did.

Since we're not yet ready to roll out strict use of version numbers
and revision specifications, I've removed the revision specification
from dn-common-utils-dev rather than adding one to dn-common-utils.