Split cps-api library into separate package.
cps-api is the largest library that has not been split into a separate
package. And because it's used so much, it limits the effectiveness of
the earlier split of the logging library, as dn-cps-api depends on the
dn-logging package instead of dn-logging-lib.

Although no version number changes is strictly needed, as the package
dependency changes will ensure dn-cps-api-lib will be fetched when it
is needed, I've bumped the patch digit of the version number to mark
this change.

Resolves: OT-83